remove AnimationNode
