+++ cocos2d/animation/types.js
@@ -1,5 +1,4 @@
 var JS = cc.js;
-var Playable = require('./playable');
 
 var WrapModeMask = {
     Loop: 1 << 1,
@@ -96,425 +95,8 @@ WrappedInfo.prototype.set = function (info) {
     this.frameIndex = info.frameIndex;
 };
 
-/**
- * !#en The abstract interface for all playing animation.
- * !#zh 所有播放动画的抽象接口。
- * @class AnimationNodeBase
- *
- * @extends Playable
- */
-function AnimationNodeBase () {
-    Playable.call(this);
-}
-JS.extend(AnimationNodeBase, Playable);
-
-/**
- * @method update
- * @param {Number} deltaTime
- * @private
- */
-AnimationNodeBase.prototype.update = function (deltaTime) {};
-
-
-/**
- * !#en The collection and instance of playing animations.
- * !#zh 动画曲线的集合，根据当前时间计算出每条曲线的状态。
- * @class AnimationNode
- * @extends AnimationNodeBase
- *
- * @param {Animator} animator
- * @param {AnimCurve[]} [curves]
- * @param {Object} [timingInput] - This dictionary is used as a convenience for specifying the timing properties of an Animation in bulk.
- */
-function AnimationNode (animator, curves, timingInput) {
-    AnimationNodeBase.call(this);
-    cc.EventTarget.call(this);
-
-    this._firstFramePlayed = false;
-
-    this._delay = 0;
-    this._delayTime = 0;
-
-    this._wrappedInfo = new WrappedInfo();
-    this._lastWrappedInfo = null;
-
-    this._process = process;
-
-    this.animator = animator;
-
-    /**
-     * !#en The curves list.
-     * !#zh 曲线列表。
-     * @property curves
-     * @type {Object[]}
-     */
-    this.curves = curves || [];
-
-    // http://www.w3.org/TR/web-animations/#idl-def-AnimationTiming
-
-    /**
-     * !#en The start delay which represents the number of seconds from an animation's start time to the start of
-     * the active interval.
-     * !#zh 延迟多少秒播放。
-     *
-     * @property delay
-     * @type {Number}
-     * @default 0
-     */
-    this.delay = 0;
-
-    /**
-     * !#en The animation's iteration count property.
-     *
-     * A real number greater than or equal to zero (including positive infinity) representing the number of times
-     * to repeat the animation node.
-     *
-     * Values less than zero and NaN values are treated as the value 1.0 for the purpose of timing model
-     * calculations.
-     *
-     * !#zh 迭代次数，指动画播放多少次后结束, normalize time。 如 2.5（2次半）
-     *
-     * @property repeatCount
-     * @type {Number}
-     * @default 1
-     */
-    this.repeatCount = 1;
-
-    /**
-     * !#en The iteration duration of this animation in seconds. (length)
-     * !#zh 单次动画的持续时间，秒。
-     *
-     * @property duration
-     * @type {Number}
-     * @readOnly
-     */
-    this.duration = 1;
-
-    /**
-     * !#en The animation's playback speed. 1 is normal playback speed.
-     * !#zh 播放速率。
-     * @property speed
-     * @type {Number}
-     * @default: 1.0
-     */
-    this.speed = 1;
-
-    /**
-     * !#en
-     * Wrapping mode of the playing animation.
-     * Notice : dynamic change wrapMode will reset time and repeatCount property
-     * !#zh
-     * 动画循环方式。
-     * 需要注意的是，动态修改 wrapMode 时，会重置 time 以及 repeatCount
-     *
-     * @property wrapMode
-     * @type {WrapMode}
-     * @default: WrapMode.Normal
-     */
-    this.wrapMode = WrapMode.Normal;
-
-    if (timingInput) {
-        this.delay = timingInput.delay || this.delay;
-
-        var duration = timingInput.duration;
-        if (typeof duration !== 'undefined') {
-            this.duration = duration;
-        }
-
-        var speed = timingInput.speed;
-        if (typeof speed !== 'undefined') {
-            this.speed = speed;
-        }
-
-        //
-        var wrapMode = timingInput.wrapMode;
-        if (typeof wrapMode !== 'undefined') {
-            var isEnum = typeof wrapMode === 'number';
-            if (isEnum) {
-                this.wrapMode = wrapMode;
-            }
-            else {
-                this.wrapMode = WrapMode[wrapMode];
-            }
-        }
-
-        var repeatCount = timingInput.repeatCount;
-        if (typeof repeatCount !== 'undefined') {
-            this.repeatCount = repeatCount;
-        }
-        else if (this.wrapMode & WrapModeMask.Loop) {
-            this.repeatCount = Infinity;
-        }
-    }
-
-    /**
-     * !#en The current time of this animation in seconds.
-     * !#zh 动画当前的时间，秒。
-     * @property time
-     * @type {Number}
-     * @default 0
-     */
-    this.time = 0;
-}
-JS.extend(AnimationNode, AnimationNodeBase);
-
-var proto = AnimationNode.prototype;
-
-function process () {
-    // sample
-    var info = this.sample();
-
-    var cache = this._hasListenerCache;
-    if (cache && cache['lastframe']) {
-        var lastInfo;
-        if (!lastInfo) {
-            lastInfo = this._lastWrappedInfo = new WrappedInfo(info);
-        }
-
-        if (this.repeatCount > 1 && 
-            ((info.iterations | 0) > (lastInfo.iterations | 0)) // another iterations
-            ) {
-            if ((this.wrapMode & WrapModeMask.Reverse) === WrapModeMask.Reverse) {
-                if (lastInfo.direction < 0) {
-                    this.emit('lastframe', this);
-                }
-            }
-            else {
-                if (lastInfo.direction > 0) {
-                    this.emit('lastframe', this);
-                }
-            }
-        }
-
-        lastInfo.set(info);
-    }
-
-    if (info.stopped) {
-        this.stop();
-        this.emit('finished', this);
-    }
-};
-
-function simpleProcess () {
-    var time = this.time;
-    var duration = this.duration;
-
-    if (time > duration) {
-        time = time % duration;
-        if (time === 0) time = duration;
-    }
-    else if (time < 0) {
-        time = time % duration;
-        if (time !== 0 ) time += duration;
-    }
-
-    var ratio = time / duration;
-
-    var curves = this.curves;
-    for (var i = 0, len = curves.length; i < len; i++) {
-        var curve = curves[i];
-        curve.sample(time, ratio, this);
-    }
-
-    var cache = this._hasListenerCache;
-    if (cache && cache['lastframe']) {
-        var currentIterations = time > 0 ? (time / duration) : -(time / duration);
-        
-        var lastIterations = this._lastIterations;
-        if (lastIterations === undefined) {
-            lastIterations = this._lastIterations = currentIterations;
-        }
-
-        if ((currentIterations | 0) > (lastIterations | 0)) {
-            this.emit('lastframe', this);
-        }
-
-        this._lastIterations = currentIterations;
-    }
-}
-
-proto.update = function (delta) {
-    // calculate delay time
-
-    if (this._delayTime > 0) {
-        this._delayTime -= delta;
-        if (this._delayTime > 0) {
-            // still waiting
-            return;
-        }
-    }
-
-    // make first frame perfect
-
-    //var playPerfectFirstFrame = (this.time === 0);
-    if (this._firstFramePlayed) {
-        this.time += (delta * this.speed);
-    }
-    else {
-        this._firstFramePlayed = true;
-    }
-
-    this._process();
-};
-
-proto._needRevers = function (currentIterations) {
-    var wrapMode = this.wrapMode;
-    var needRevers = false;
-
-    if ((wrapMode & WrapModeMask.PingPong) === WrapModeMask.PingPong) {
-        var isEnd = currentIterations - (currentIterations | 0) === 0;
-        if (isEnd && (currentIterations > 0)) {
-            currentIterations -= 1;
-        }
-
-        var isOddIteration = currentIterations & 1;
-        if (isOddIteration) {
-            needRevers = !needRevers;
-        }
-    }
-    if ((wrapMode & WrapModeMask.Reverse) === WrapModeMask.Reverse) {
-        needRevers = !needRevers;
-    }
-    return needRevers;
-};
-
-proto.getWrappedInfo = function (time, info) {
-    info = info || new WrappedInfo();
-    
-    var stopped = false;
-    var duration = this.duration;
-    var repeatCount = this.repeatCount;
-
-    var currentIterations = time > 0 ? (time / duration) : -(time / duration);
-    if (currentIterations >= repeatCount) {
-        currentIterations = repeatCount;
-
-        stopped = true;
-        var tempRatio = repeatCount - (repeatCount | 0);
-        if (tempRatio === 0) {
-            tempRatio = 1;  // 如果播放过，动画不复位
-        }
-        time = tempRatio * duration * (time > 0 ? 1 : -1);
-    }
-
-    if (time > duration) {
-        var tempTime = time % duration;
-        time = tempTime === 0 ? duration : tempTime;
-    }
-    else if (time < 0) {
-        time = time % duration;
-        if (time !== 0 ) time += duration;
-    }
-
-    var needRevers = false;
-    var shouldWrap = this._wrapMode & WrapModeMask.ShouldWrap;
-    if (shouldWrap) {
-        needRevers = this._needRevers(currentIterations);
-    }
-
-    var direction = needRevers ? -1 : 1;
-    if (this.speed < 0) {
-        direction *= -1;
-    }
-
-    // calculate wrapped time
-    if (shouldWrap && needRevers) {
-        time = duration - time;
-    }
-
-    info.ratio = time / duration;
-    info.time = time;
-    info.direction = direction;
-    info.stopped = stopped;
-    info.iterations = currentIterations;
-
-    return info;
-};
-
-proto.sample = function () {
-    var info = this.getWrappedInfo(this.time, this._wrappedInfo);
-    var curves = this.curves;
-    for (var i = 0, len = curves.length; i < len; i++) {
-        var curve = curves[i];
-        curve.sample(info.time, info.ratio, this);
-    }
-
-    return info;
-};
-
-proto.onStop = function () {
-    this.emit('stop', this);
-};
-
-proto.onPlay = function () {
-    this._delayTime = this._delay
-    this.emit('play', this);
-};
-
-proto.onPause = function () {
-    this.emit('pause', this);
-};
-
-proto.onResume = function () {
-    this.emit('resume', this);
-};
-
-JS.getset(proto, 'wrapMode',
-    function () {
-        return this._wrapMode;
-    },
-    function (value) {
-        this._wrapMode = value;
-
-        if (CC_EDITOR) return;
-
-        // dynamic change wrapMode will need reset time to 0
-        this.time = 0;
-
-        if (value & WrapModeMask.Loop) {
-            this.repeatCount = Infinity;
-        }
-        else {
-            this.repeatCount = 1;
-        }
-        
-    }
-);
-
-JS.getset(proto, 'repeatCount',
-    function () {
-        return this._repeatCount;
-    },
-    function (value) {
-        this._repeatCount = value;
-        
-        var shouldWrap = this._wrapMode & WrapModeMask.ShouldWrap;
-        var reverse = (this.wrapMode & WrapModeMask.Reverse) === WrapModeMask.Reverse;
-        if (value === Infinity && !shouldWrap && !reverse) {
-            this._process = simpleProcess;
-        }
-        else {
-            this._process = process;
-        }
-    }
-);
-
-JS.getset(proto, 'delay', 
-    function () {
-        return this._delay;
-    },
-    function (value) {
-        this._delayTime = this._delay = value;
-    }
-);
-
-cc.js.mixin(proto, cc.EventTarget.prototype);
-
-cc.AnimationNode = AnimationNode;
-
 module.exports = {
     WrapModeMask,
     WrapMode,
-    AnimationNode,
     WrappedInfo
 };
